+++ components.ts
@@ -1,0 +1,5 @@
+/**
+ * @vowel.to/client/components - Web component exports for vowel.
+ */
+
+export * from './web-component.js';